stream: Call read(0) on 'readable' emission
This is so that, in the case where the user does not call read()
to consume all the bytes and get below the low water mark, we can
still check and try to call _read if appropriate.

This is especially important for hitting the EOF so that we know
to tear down TCP sockets.